Update AddOrgModal.js (#606)
* Update AddOrgModal.js

Centered form, centered modal, added effect, added click anywhere close

* Finalized AddOrgModal

Created folder to house component. Separated style and index and included CountrySelect component. Add scroll bar to modal. Changed TextField inputs to 'small' in all display sizes.
